Export utils context consumer
I've created a set of Picker components for a project using `createAutoCorrectedDatePipe` from text-mask-addons. To make this work seamlessly I'm using the `disablePast`, `disableFuture`, `minDate`, and `maxDate` props to set the correct `minYear` and `maxYear` settings for the pipe.

These components will be part of an internal component library that we are currently developing. They should not be locked to a specific date library, and instead always use the same as material-ui-pickers uses.

My solution for now has been to provide the same utils object in my own context so that I can consume it. But this creates internal dependencies that that I'd like to avoid. 

Using the context consumer from material-ui-pickers would ensure that I was always using the same date utils without duplicating the context myself.